+++ e2e/e2etests/e2etests.go
@@ -82,6 +82,7 @@ const (
 	TestBitcoinWithdrawP2SHName           = "bitcoin_withdraw_p2sh"
 	TestBitcoinWithdrawInvalidAddressName = "bitcoin_withdraw_invalid"
 	TestBitcoinWithdrawRestrictedName     = "bitcoin_withdraw_restricted"
+	TestExtractBitcoinInscriptionMemoName = "bitcoin_memo_from_inscription"
 
 	/*
 	 Application tests
@@ -131,7 +132,9 @@ const (
 	TestV2ETHDepositAndCallRevertName            = "v2_eth_deposit_and_call_revert"
 	TestV2ETHDepositAndCallRevertWithCallName    = "v2_eth_deposit_and_call_revert_with_call"
 	TestV2ETHWithdrawName                        = "v2_eth_withdraw"
+	TestV2ETHWithdrawAndArbitraryCallName        = "v2_eth_withdraw_and_arbitrary_call"
 	TestV2ETHWithdrawAndCallName                 = "v2_eth_withdraw_and_call"
+	TestV2ETHWithdrawAndCallThroughContractName  = "v2_eth_withdraw_and_call_through_contract"
 	TestV2ETHWithdrawAndCallRevertName           = "v2_eth_withdraw_and_call_revert"
 	TestV2ETHWithdrawAndCallRevertWithCallName   = "v2_eth_withdraw_and_call_revert_with_call"
 	TestV2ERC20DepositName                       = "v2_erc20_deposit"
@@ -142,7 +145,9 @@ const (
 	TestV2ERC20WithdrawAndCallName               = "v2_erc20_withdraw_and_call"
 	TestV2ERC20WithdrawAndCallRevertName         = "v2_erc20_withdraw_and_call_revert"
 	TestV2ERC20WithdrawAndCallRevertWithCallName = "v2_erc20_withdraw_and_call_revert_with_call"
+	TestV2ZEVMToEVMArbitraryCallName             = "v2_zevm_to_evm_arbitrary_call"
 	TestV2ZEVMToEVMCallName                      = "v2_zevm_to_evm_call"
+	TestV2ZEVMToEVMCallThroughContractName       = "v2_zevm_to_evm_call_through_contract"
 	TestV2EVMToZEVMCallName                      = "v2_evm_to_zevm_call"
 
 	/*
@@ -160,6 +165,9 @@ const (
 	TestPrecompilesPrototypeThroughContractName = "precompile_contracts_prototype_through_contract"
 	TestPrecompilesStakingName                  = "precompile_contracts_staking"
 	TestPrecompilesStakingThroughContractName   = "precompile_contracts_staking_through_contract"
+	TestPrecompilesBankName                     = "precompile_contracts_bank"
+	TestPrecompilesBankFailName                 = "precompile_contracts_bank_fail"
+	TestPrecompilesBankThroughContractName      = "precompile_contracts_bank_through_contract"
 )
 
 // AllE2ETests is an ordered list of all e2e tests
@@ -444,6 +452,13 @@ var AllE2ETests = []runner.E2ETest{
 	/*
 	 Bitcoin tests
 	*/
+	runner.NewE2ETest(
+		TestExtractBitcoinInscriptionMemoName,
+		"extract memo from BTC inscription", []runner.ArgDefinition{
+			{Description: "amount in btc", DefaultValue: "0.1"},
+		},
+		TestExtractBitcoinInscriptionMemo,
+	),
 	runner.NewE2ETest(
 		TestBitcoinDepositName,
 		"deposit Bitcoin into ZEVM",
@@ -731,12 +746,28 @@ var AllE2ETests = []runner.E2ETest{
 		TestV2ETHWithdraw,
 	),
 	runner.NewE2ETest(
-		TestV2ETHWithdrawAndCallName,
+		TestV2ETHWithdrawAndArbitraryCallName,
 		"withdraw Ether from ZEVM and call a contract using V2 contract",
 		[]runner.ArgDefinition{
 			{Description: "amount in wei", DefaultValue: "100000"},
 		},
+		TestV2ETHWithdrawAndArbitraryCall,
+	),
+	runner.NewE2ETest(
+		TestV2ETHWithdrawAndCallName,
+		"withdraw Ether from ZEVM call a contract using V2 contract",
+		[]runner.ArgDefinition{
+			{Description: "amount in wei", DefaultValue: "100000"},
+		},
 		TestV2ETHWithdrawAndCall,
+	),
+	runner.NewE2ETest(
+		TestV2ETHWithdrawAndCallThroughContractName,
+		"withdraw Ether from ZEVM call a contract using V2 contract through intermediary contract",
+		[]runner.ArgDefinition{
+			{Description: "amount in wei", DefaultValue: "100000"},
+		},
+		TestV2ETHWithdrawAndCallThroughContract,
 	),
 	runner.NewE2ETest(
 		TestV2ETHWithdrawAndCallRevertName,
@@ -819,10 +850,22 @@ var AllE2ETests = []runner.E2ETest{
 		TestV2ERC20WithdrawAndCallRevertWithCall,
 	),
 	runner.NewE2ETest(
+		TestV2ZEVMToEVMArbitraryCallName,
+		"zevm -> evm call using V2 contract",
+		[]runner.ArgDefinition{},
+		TestV2ZEVMToEVMArbitraryCall,
+	),
+	runner.NewE2ETest(
 		TestV2ZEVMToEVMCallName,
 		"zevm -> evm call using V2 contract",
 		[]runner.ArgDefinition{},
 		TestV2ZEVMToEVMCall,
+	),
+	runner.NewE2ETest(
+		TestV2ZEVMToEVMCallThroughContractName,
+		"zevm -> evm call using V2 contract through intermediary contract",
+		[]runner.ArgDefinition{},
+		TestV2ZEVMToEVMCallThroughContract,
 	),
 	runner.NewE2ETest(
 		TestV2EVMToZEVMCallName,
@@ -885,5 +928,23 @@ var AllE2ETests = []runner.E2ETest{
 		"test stateful precompiled contracts staking through contract",
 		[]runner.ArgDefinition{},
 		TestPrecompilesStakingThroughContract,
+	),
+	runner.NewE2ETest(
+		TestPrecompilesBankName,
+		"test stateful precompiled contracts bank with ZRC20 tokens",
+		[]runner.ArgDefinition{},
+		TestPrecompilesBank,
+	),
+	runner.NewE2ETest(
+		TestPrecompilesBankFailName,
+		"test stateful precompiled contracts bank with non ZRC20 tokens",
+		[]runner.ArgDefinition{},
+		TestPrecompilesBankNonZRC20,
+	),
+	runner.NewE2ETest(
+		TestPrecompilesBankThroughContractName,
+		"test stateful precompiled contracts bank through contract",
+		[]runner.ArgDefinition{},
+		TestPrecompilesBankThroughContract,
 	),
 }