+++ e2e/e2etests/test_v2_erc20_deposit_and_call_revert_with_call.go
@@ -3,6 +3,7 @@ package e2etests
 import (
 	"math/big"
 
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/stretchr/testify/require"
 	"github.com/zeta-chain/protocol-contracts/v2/pkg/gatewayevm.sol"
 
@@ -38,4 +39,12 @@ func TestV2ERC20DepositAndCallRevertWithCall(r *runner.E2ERunner, args []string)
 
 	// check the payload was received on the contract
 	r.AssertTestDAppEVMCalled(true, payloadMessageDepositOnRevertERC20, big.NewInt(0))
+
+	// check expected sender was used
+	senderForMsg, err := r.TestDAppV2EVM.SenderWithMessage(
+		&bind.CallOpts{},
+		[]byte(payloadMessageDepositOnRevertERC20),
+	)
+	require.NoError(r, err)
+	require.Equal(r, r.EVMAuth.From, senderForMsg)
 }