+++ x/fungible/keeper/v2_deposits.go
@@ -49,6 +49,7 @@ func (k Keeper) ProcessV2Deposit(
 // ProcessV2RevertDeposit handles a revert deposit from an inbound tx with protocol version 2
 func (k Keeper) ProcessV2RevertDeposit(
 	ctx sdk.Context,
+	inboundSender string,
 	amount *big.Int,
 	chainID int64,
 	coinType coin.CoinType,
@@ -74,7 +75,7 @@ func (k Keeper) ProcessV2RevertDeposit(
 
 		if callOnRevert {
 			// no asset, call simple revert
-			_, err := k.CallExecuteRevert(ctx, zrc20Addr, amount, revertAddress, revertMessage)
+			_, err := k.CallExecuteRevert(ctx, inboundSender, zrc20Addr, amount, revertAddress, revertMessage)
 			return err
 		} else {
 			// no asset, no call, do nothing
@@ -87,6 +88,7 @@ func (k Keeper) ProcessV2RevertDeposit(
 			// revert with a ZRC20 asset
 			_, err := k.CallDepositAndRevert(
 				ctx,
+				inboundSender,
 				zrc20Addr,
 				amount,
 				revertAddress,