+++ zetaclient/chains/bitcoin/signer/signer.go
@@ -154,7 +154,10 @@ func (signer *Signer) AddWithdrawTxOutputs(
 	}
 
 	// 1st output: the nonce-mark btc to TSS self
-	tssAddrP2WPKH := signer.TSS().BTCAddressWitnessPubkeyHash()
+	tssAddrP2WPKH, err := signer.TSS().BTCAddress(signer.Chain().ChainId)
+	if err != nil {
+		return err
+	}
 	payToSelfScript, err := bitcoin.PayToAddrScript(tssAddrP2WPKH)
 	if err != nil {
 		return err
@@ -381,7 +384,7 @@ func (signer *Signer) TryProcessOutbound(
 	}
 
 	// get size limit and gas price
-	sizelimit := params.GasLimit
+	sizelimit := params.CallOptions.GasLimit
 	gasprice, ok := new(big.Int).SetString(params.GasPrice, 10)
 	if !ok || gasprice.Cmp(big.NewInt(0)) < 0 {
 		logger.Error().Msgf("cannot convert gas price  %s ", params.GasPrice)